Add dispatch table entries required by RenderScript HIDL implementation.
  - Add several missing entries to the dispatch table. Dispatch table is
    now used for switching between HIDL implementation and our CPU
    fallback, instead of  only accessing the APIs exposed by RS
    support library and RS NDK library.

Bug: 34396220
Test: mm

+++ support/jni/android_renderscript_RenderScript.cpp
@@ -989,7 +989,7 @@ nElementGetSubElements(JNIEnv *_env, jobject _this, jlong con, jlong id,
 
     uintptr_t *ids = (uintptr_t *)malloc(dataSize * sizeof(uintptr_t));
     const char **names = (const char **)malloc((uint32_t)dataSize * sizeof(const char *));
-    uint32_t *arraySizes = (uint32_t *)malloc((uint32_t)dataSize * sizeof(uint32_t));
+    size_t *arraySizes = (size_t *)malloc(dataSize * sizeof(size_t));
 
     dispatchTab.ElementGetSubElements((RsContext)con, (RsElement)id, ids, names, arraySizes,
                                       (uint32_t)dataSize);
